animation restructure
