+++ packages/gamut/src/Pagination/AnimatedPaginationButtons.tsx
@@ -1,5 +1,5 @@
-import { createAnimatedFadeButton, wrapWithSlideAnimation } from './animations';
 import { PaginationButton } from './PaginationButton';
+import { createAnimatedFadeButton, wrapWithSlideAnimation } from './utils';
 
 export const AnimatedFadeButton = createAnimatedFadeButton(PaginationButton);
 export const AnimatedSlideButton = wrapWithSlideAnimation(PaginationButton);
